Day 4: Forklift Maneouverability (and adding a new day template)
I came up with my solution to part 1 pretty quickly. It's not very
clever, and neither was my solution to part 2. All I did was extract the
logic for P1 into a function and then apply that function over and over
until it stopped changing. Either way, it ran fast and gave the right
answer, so how clever it is doesn't really matter. Good puzzle.