Day 5: Valid IDs, Again
Part 1 was a rather simple test of checking whether a number is in a
range, no biggie. Part 2 was more interesting, as simply going through
each range and counting without duplicates would've taken rather a while
and a lot of memory (don't ask how I know). Instead, the answer is just
to simplify ranges down as much as possible and total the length of
those instead.