Day 6: Cephalopodian Maths
This was a fun one, and I liked the part 2 twist. Part 1 had a lot of
opportunites for fun list comprehensions, but otherwise was relatively
simple. Part 2 had way more obstacles, however. Firstly was correctly
partitioning each row of operands. Then was getting vertically read
numbers. But once the parsing was sorted, the calculation itself was the
same as part 1. Overall, I liked this day.