Day 8: Junction Boxes, AKA the one that broke me
Good Lord, what the hell was this? I hadn't heard of Kruskal's algorithm
before, so I initially misunderstood this puzzle dearly. Then, once I
did understand it, I implemented a very slow and naiive solution, but it
did eventually work. Then I tried to understand Kruskal's and DSUs, both
of which gave me a headache. But, we have eventually got here, and it is
an admittedly neat DSA combo. The largest downside to all this though is
just how long it took for me to understand. At least there's fewer
puzzles this year.